add redis for fast api response

+++ src/app/(server)/api/notification/delete/route.ts
@@ -1,11 +1,22 @@
+import { redis } from "@/lib/redis";
+import dbConnect from "@/mongodb/connectDB";
 import notificationModel from "@/mongodb/notifications.model";
+import { ProtectRoute } from "@/services/protectRoute";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(req: NextRequest) {
     try {
+        await dbConnect();
+        
+        const isValid = ProtectRoute(req);
+
+        if (!isValid) return NextResponse.json({ error: "Unauthorized User" }, { status: 401 });
+
         const { deleteNotificationId } = await req.json();
 
         const deletedNotification = await notificationModel.findByIdAndDelete(deleteNotificationId);
+
+        await redis.del(deletedNotification.userEmail);
 
         return NextResponse.json({ success: true, deletedNotification: deletedNotification }, { status: 200 })
     } catch (err) {
